+++ client/battle/BattleEffectsController.cpp
@@ -18,9 +18,9 @@
 #include "BattleStacksController.h"
 #include "BattleRenderer.h"
 
-#include "../CMusicHandler.h"
 #include "../CGameInfo.h"
 #include "../CPlayerInterface.h"
+#include "../media/ISoundPlayer.h"
 #include "../render/Canvas.h"
 #include "../render/CAnimation.h"
 #include "../render/Graphics.h"
