+++ client/lobby/OptionsTab.cpp
@@ -14,12 +14,12 @@
 
 #include "../CGameInfo.h"
 #include "../CServerHandler.h"
-#include "../CMusicHandler.h"
 #include "../gui/CGuiHandler.h"
 #include "../gui/Shortcut.h"
 #include "../gui/WindowHandler.h"
 #include "../render/Graphics.h"
 #include "../render/IFont.h"
+#include "../media/ISoundPlayer.h"
 #include "../widgets/CComponent.h"
 #include "../widgets/ComboBox.h"
 #include "../widgets/CTextInput.h"
@@ -38,6 +38,7 @@
 #include "../../lib/networkPacks/PacksForLobby.h"
 #include "../../lib/CGeneralTextHandler.h"
 #include "../../lib/CArtHandler.h"
+#include "../../lib/CConfigHandler.h"
 #include "../../lib/CTownHandler.h"
 #include "../../lib/CHeroHandler.h"
 #include "../../lib/mapping/CMapInfo.h"
